python-multi-tenancy-first/pyproject.toml
[project]
name = "python-multi-tenancy-first"
version = "0.1.0"
description = "Subdomain-based multi-tenant API using FastAPI + SQLAlchemy 2.0 + Postgres RLS"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi (>=0.110,<1.0)",
    "uvicorn (>=0.29,<1.0)",
    "sqlalchemy (>=2.0,<3.0)",
    "psycopg2-binary (>=2.9,<3.0)",
    "PyJWT (>=2.8,<3.0)",
    "passlib[bcrypt] (>=1.7,<2.0)",
    "bcrypt (>=4.0,<5.0)",
    "python-dotenv (>=1.0,<2.0)",
]

[tool.poetry]
packages = [{include = "multi_tenancy", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
httpx = "^0.28"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
